# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products": {
      "post": {
        "operationId": "ProductsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Product"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      },
      "get": {
        "operationId": "ProductsController_findAll_v1",
        "parameters": [
          {
            "name": "current",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seller_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "price_min",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "price_max",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        },
                        "result": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Product"
                          }
                        }
                      },
                      "required": [
                        "meta",
                        "result"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/mine": {
      "get": {
        "operationId": "ProductsController_findMine_v1",
        "parameters": [
          {
            "name": "current",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        },
                        "result": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Product"
                          }
                        }
                      },
                      "required": [
                        "meta",
                        "result"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "summary": "Hàng của CHÍNH người đang đăng nhập — đủ mọi trạng thái.\n\nĐặt TRƯỚC `@Get(':id')`: Nest so khớp route theo thứ tự khai báo, nên nếu\nđứng sau thì `/products/mine` sẽ bị `:id` nuốt và `+'mine'` ra `NaN`.\n\n`sellerId` lấy từ TOKEN, không từ query — người bán chỉ xem được hàng của\nchính mình. Đó cũng là lý do đường này tồn tại: `GET /products` là công\nkhai nên không phân biệt được ai đang hỏi, và từ 25/09 nó lọc cứng\n`status = active` để tin nháp / bị từ chối duyệt không bày ra cho người lạ.",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Product"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Product"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{id}/stock": {
      "patch": {
        "operationId": "ProductsController_updateStock_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStockDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Product"
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
...
    "/api/v1/interactions/product/{productId}": {
      "get": {
        "operationId": "InteractionsController_findByProduct_v1",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currentPage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        },
                        "result": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Review"
                          }
                        }
                      },
                      "required": [
                        "meta",
                        "result"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Interactions"
        ]
      }
    },
...
    "/api/v1/products/{id}/related": {
      "get": {
        "operationId": "RecommendationsController_related_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "summary": "\"Người mua món này cũng mua\" — trang chi tiết sản phẩm.\n\nCÔNG KHAI, không cần đăng nhập: khách chưa có tài khoản vẫn xem trang sản\nphẩm, và đó chính là lúc gợi ý có giá trị nhất. Vì công khai nên nó chỉ\ntrả về sản phẩm `active` còn hàng — không rò gì về người mua.",
        "tags": [
          "Recommendations"
        ]
      }
    },
    "/api/v1/recommendations/me": {
      "get": {
        "operationId": "MyRecommendationsController_forMe_v1",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    }
                  },
                  "required": [
                    "statusCode",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "MyRecommendations"
        ]
      }
    }
  },
  "info": {
    "title": "Zoldify API",
    "description": "API của sàn mua bán đồ cũ Zoldify. Web (Next.js) và app (React Native) dùng chung hợp đồng này.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "current": {
            "type": "number",
            "description": "Trang hiện tại, bắt đầu từ 1"
          },
          "pageSize": {
            "type": "number",
            "description": "Số bản ghi mỗi trang"
          },
          "pages": {
            "type": "number",
            "description": "Tổng số trang"
          },
          "total": {
            "type": "number",
            "description": "Tổng số bản ghi khớp điều kiện lọc"
          }
        },
        "required": [
          "current",
          "pageSize",
          "pages",
          "total"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "full_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "role": {
            "enum": [
              "buyer",
              "seller",
              "admin",
              "moderator"
            ],
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          },
          "is_locked": {
            "type": "boolean"
          },
          "last_seen": {
            "format": "date-time",
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "full_name",
          "email",
          "phone_number",
          "role",
          "avatar",
          "email_verified",
          "is_locked",
          "last_seen",
          "gender",
          "created_at"
        ]
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "name_en": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          }
        },
        "required": [
          "id",
          "name",
          "name_en",
          "description",
          "slug",
          "image",
          "is_active",
          "created_at",
          "updated_at",
          "products"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "currency": {
            "type": "string",
            "description": "4.1. Đơn vị tiền của `price` — mã ISO 4217 ('VND', 'USD', 'JPY'…).\n\nVÌ SAO CẦN: trước cột này, mọi con số trong hệ thống đều NGẦM hiểu là đồng\nViệt Nam, và không chỗ nào trong mã nói ra điều đó. Giao diện tiếng Anh vì\nvậy vẫn in \"1.890.000 ₫\" — không phải lỗi định dạng, mà là vì thật sự không\ncó dữ liệu nào cho nó biết số đó là tiền gì.\n\n⚠️ CỘT NÀY CHỈ LÀM ĐÚNG PHẦN HIỂN THỊ. Nó KHÔNG khiến sàn giao dịch được\nxuyên tiền tệ. Muốn thế còn cần: nguồn tỉ giá, thời điểm chốt giá, và một\nsổ cái nhiều tiền tệ. Sổ cái hiện tại (ledger_accounts) kiểm bất biến\n`sum(entries) = 0`, mà bất biến đó chỉ có nghĩa TRONG CÙNG một tiền tệ —\ntrộn hai loại tiền vào một sổ là cách làm hỏng sổ sách mà không ai thấy.\nNên tới khi có thiết kế đó, mỗi sàn vẫn nên chạy một tiền tệ."
          },
          "stock": {
            "type": "number"
          },
          "image": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "spec": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "condition": {
            "type": "string"
          },
          "is_freeship": {
            "type": "boolean"
          },
          "sold_count": {
            "type": "number"
          },
          "view_count": {
            "type": "number"
          },
          "status": {
            "enum": [
              "draft",
              "pending",
              "active",
              "sold",
              "rejected"
            ],
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "seller": {
            "$ref": "#/components/schemas/User"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "price",
          "currency",
          "stock",
          "image",
          "brand",
          "spec",
          "images",
          "condition",
          "is_freeship",
          "sold_count",
          "view_count",
          "status",
          "category",
          "seller",
          "created_at",
          "updated_at"
        ]
      },
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "currency": {
            "type": "string",
            "description": "Mã ISO 4217, 3 chữ hoa. Bỏ trống thì entity mặc định 'VND'.\n\nChỉ kiểm HÌNH DẠNG chứ không kiểm mã có thật: danh sách ISO 4217 thay đổi\ntheo thời gian và giữ một bản chép tay ở đây thì sớm muộn cũng lạc hậu.\nMã lạ lọt qua sẽ hiện ra ở phần định dạng phía frontend (Intl ném lỗi và\nrơi về in số kèm mã), chứ không làm hỏng dữ liệu.",
            "pattern": "^[A-Z]{3}$"
          },
          "image": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "category_id": {
            "type": "number"
          },
          "brand": {
            "type": "string"
          },
          "spec": {
            "type": "string"
          },
          "stock": {
            "type": "number"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "condition": {
            "type": "string"
          },
          "is_freeship": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "price",
          "category_id"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateStockDto": {
        "type": "object",
        "properties": {
          "stock": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "stock"
        ]
      },
...
      "Review": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          },
          "rating": {
            "type": "number"
          },
          "comment": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user",
          "product",
          "order",
          "rating",
          "comment",
          "images",
          "created_at",
          "updated_at"
        ]
      },
...
    }
  }
}